Uses the initalized Pluggable plugin if available, also moves the file one level up.
